survey page now returning data from the DB, with survey ID currently hardcoded
